Extract navigation config to a separate file
